requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.1
grpcio==1.62.1
grpcio-tools==1.62.1
